+++ CHANGELOG.md
@@ -5,12 +5,13 @@ A more detailed list of changes is available in the corresponding milestones for
 ## 0.7.7 (2019-Jun-17)
 ### Note-worthy code changes
   - **[com.google.fonts/check/family/has_license]:** only run check if the fonts are in a google/fonts repo.
+  - **[com.google.fonts/check/vendor_id]:** accept NULL-padding on vendor-IDs (issue #2548)
 
 ### Bug fixes
   - fix crash on git_gfonts_ttFonts condition (return None when the font is not yet available on GitHub) (issue #2540)
 
 ### New Checks
- - **[com.google.fonts/check/metadata/broken_links]:** Ensure the copyright string in METADATA.pb files does not contain broken URLs. (issue #2550)
+  - **[com.google.fonts/check/metadata/broken_links]:** Ensure the copyright string in METADATA.pb files does not contain broken URLs. (issue #2550)
 
 
 ## 0.7.6 (2019-Jun-10)